Feature complete.
- Fixed the "deadlock" issue while invoking the python process from C# with redirected I/O
- Python scripts were modified to produce no other output than the desired one.

When python detects there is no TTS attached to the process, it enters idle mode and performs no process. Therefore, when C# calls the process it does not detect the user input console and enters in sleep mode stopping any further processing. To prevent this the python process must be called with the -i argument.